+++ lib-mal/src/lib.rs
@@ -3,12 +3,22 @@ mod test;
 
 pub mod model;
 
-use model::{AnimeDetails, AnimeList, ForumBoards, ListStatus, fields::AnimeField, options::{RankingType, Season, Params}};
+use model::{
+    fields::AnimeField,
+    options::{Params, RankingType, Season},
+    AnimeDetails, AnimeList, ForumBoards, ForumTopics, ListStatus, TopicDetails,
+};
 
 use pkce;
 use reqwest::{Method, StatusCode};
 use serde::{Deserialize, Serialize};
-use std::{fs::{self, File}, io::Write, path::PathBuf, str, time::SystemTime};
+use std::{
+    fs::{self, File},
+    io::Write,
+    path::PathBuf,
+    str,
+    time::SystemTime,
+};
 use tiny_http::{Response, Server};
 
 pub struct MALClient {
@@ -25,8 +35,8 @@ impl MALClient {
         let client = reqwest::Client::new();
         let mut will_cache = caching;
         let mut n_a = false;
-        let dir = if let Some(d) = cache_dir{
-           d
+        let dir = if let Some(d) = cache_dir {
+            d
         } else {
             println!("No cache directory was provided, disabling caching");
             will_cache = false;
@@ -150,8 +160,8 @@ impl MALClient {
                 .as_secs(),
         };
         if self.caching {
-            let mut f = File::create(self.dirs.join("tokens.json"))
-                .expect("Unable to create token file");
+            let mut f =
+                File::create(self.dirs.join("tokens.json")).expect("Unable to create token file");
             f.write_all(serde_json::to_string(&tjson).unwrap().as_bytes())
                 .expect("Unable to write tokens");
         }
@@ -159,27 +169,46 @@ impl MALClient {
 
     ///Sends a get request to the specified URL with the appropriate auth header
     async fn do_request(&self, url: String) -> Result<String, String> {
-        match self.client.get(url).bearer_auth(&self.access_token).send().await {
+        match self
+            .client
+            .get(url)
+            .bearer_auth(&self.access_token)
+            .send()
+            .await
+        {
             Ok(res) => Ok(res.text().await.unwrap()),
-            Err(e) => Err(format!("{}", e))
+            Err(e) => Err(format!("{}", e)),
         }
     }
 
-    ///Sends a put request to the specified URL with the appropriate auth header and 
+    ///Sends a put request to the specified URL with the appropriate auth header and
     ///form encoded parameters
-    async fn do_request_forms(&self, url: String, params: Vec<(&str, String)>) -> Result<String, String> {
-        match self.client.put(url).bearer_auth(&self.access_token).form(&params).send().await {
+    async fn do_request_forms(
+        &self,
+        url: String,
+        params: Vec<(&str, String)>,
+    ) -> Result<String, String> {
+        match self
+            .client
+            .put(url)
+            .bearer_auth(&self.access_token)
+            .form(&params)
+            .send()
+            .await
+        {
             Ok(res) => Ok(res.text().await.unwrap()),
-            Err(e) => Err(format!("{}", e))
+            Err(e) => Err(format!("{}", e)),
         }
     }
 
     ///Tries to parse a JSON response string into the type provided in the `::<>` turbofish
-    fn parse_response<'a, T: Serialize + Deserialize<'a>>(&self, res: &'a str) -> Result<T, String> {
-
+    fn parse_response<'a, T: Serialize + Deserialize<'a>>(
+        &self,
+        res: &'a str,
+    ) -> Result<T, String> {
         match serde_json::from_str::<T>(res) {
             Ok(v) => Ok(v),
-            Err(e) => Err(format!("{}", e))
+            Err(e) => Err(format!("{}", e)),
         }
     }
 
@@ -188,10 +217,18 @@ impl MALClient {
     //--Anime functions--//
     ///Gets a list of anime based on the query string provided
     ///`limit` defaults to 100 if `None`
-    pub async fn get_anime_list(&self, query: &str, limit: Option<u8>) -> Result<AnimeList, String> {
-       let url = format!("https://api.myanimelist.net/v2/anime?q={}&limit={}", query, limit.unwrap_or(100));
-       let res = self.do_request(url).await?;
-       self.parse_response(&res)
+    pub async fn get_anime_list(
+        &self,
+        query: &str,
+        limit: Option<u8>,
+    ) -> Result<AnimeList, String> {
+        let url = format!(
+            "https://api.myanimelist.net/v2/anime?q={}&limit={}",
+            query,
+            limit.unwrap_or(100)
+        );
+        let res = self.do_request(url).await?;
+        self.parse_response(&res)
     }
 
     ///Gets the deatils for an anime by the show's ID.
@@ -202,15 +239,22 @@ impl MALClient {
         &self,
         id: &u32,
         fields: Option<Vec<AnimeField>>,
-    ) -> Result<AnimeDetails, String> { 
+    ) -> Result<AnimeDetails, String> {
         let url = if let Some(f) = fields {
             format!(
                 "https://api.myanimelist.net/v2/anime/{}?fields={}",
                 id,
-                f.iter().map(|v| v.to_string()).collect::<Vec<String>>().join(",")
+                f.iter()
+                    .map(|v| v.to_string())
+                    .collect::<Vec<String>>()
+                    .join(",")
             )
         } else {
-            format!("https://api.myanimelist.net/v2/anime/{}?fields={}", id, AnimeField::ALL)
+            format!(
+                "https://api.myanimelist.net/v2/anime/{}?fields={}",
+                id,
+                AnimeField::ALL
+            )
         };
         match self
             .client
@@ -224,11 +268,19 @@ impl MALClient {
         }
     }
 
-    ///Gets a list of anime ranked by `RankingType` 
+    ///Gets a list of anime ranked by `RankingType`
     ///
     ///`limit` defaults to the max of 100 when `None`
-    pub async fn get_anime_ranking(&self, ranking_type: RankingType, limit: Option<u8>) -> Result<AnimeList, String> {
-        let url = format!("https://api.myanimelist.net/v2/anime/ranking?ranking_type={}&limit={}", ranking_type, limit.unwrap_or(100));
+    pub async fn get_anime_ranking(
+        &self,
+        ranking_type: RankingType,
+        limit: Option<u8>,
+    ) -> Result<AnimeList, String> {
+        let url = format!(
+            "https://api.myanimelist.net/v2/anime/ranking?ranking_type={}&limit={}",
+            ranking_type,
+            limit.unwrap_or(100)
+        );
         let res = self.do_request(url).await?;
         Ok(serde_json::from_str(&res).unwrap())
     }
@@ -236,24 +288,40 @@ impl MALClient {
     ///Gets the anime for a given season in a given year
     ///
     ///`limit` defaults to the max of 100 when `None`
-    pub async fn get_seasonal_anime(&self, season: Season, year: u32, limit: Option<u8>) -> Result<AnimeList, String> {
-        let url = format!("https://api.myanimelist.net/v2/anime/season/{}/{}?limit={}", year, season, limit.unwrap_or(100)); 
+    pub async fn get_seasonal_anime(
+        &self,
+        season: Season,
+        year: u32,
+        limit: Option<u8>,
+    ) -> Result<AnimeList, String> {
+        let url = format!(
+            "https://api.myanimelist.net/v2/anime/season/{}/{}?limit={}",
+            year,
+            season,
+            limit.unwrap_or(100)
+        );
         let res = self.do_request(url).await?;
-        self.parse_response(&res) 
+        self.parse_response(&res)
     }
 
     ///Returns the suggested anime for the current user. Can return an empty list if the user has
     ///no suggestions.
     pub async fn get_suggested_anime(&self, limit: Option<u8>) -> Result<AnimeList, String> {
-        let url = format!("https://api.myanimelist.net/v2/anime/suggestions?limit={}", limit.unwrap_or(100));
+        let url = format!(
+            "https://api.myanimelist.net/v2/anime/suggestions?limit={}",
+            limit.unwrap_or(100)
+        );
         let res = self.do_request(url).await?;
         self.parse_response(&res)
     }
 
     //--User anime list functions--//
 
-    
-    pub async fn update_user_anime_status<T: Params>(&self, id: u32, update: T) -> Result<ListStatus, String> {
+    pub async fn update_user_anime_status<T: Params>(
+        &self,
+        id: u32,
+        update: T,
+    ) -> Result<ListStatus, String> {
         let params = update.get_params();
         let url = format!("https://api.myanimelist.net/v2/anime/{}/my_list_status", id);
         let res = self.do_request_forms(url, params).await?;
@@ -261,44 +329,95 @@ impl MALClient {
         self.parse_response(&res)
     }
 
-
     ///Returns the user's full anime list as an `AnimeList` struct.
     ///If the request fails for any reason, an `Err` object with a string describing the error is returned instead
     pub async fn get_user_anime_list(&self) -> Result<AnimeList, String> {
         let url = "https://api.myanimelist.net/v2/users/@me/animelist?fields=list_status&limit=4";
         let res = self.do_request(url.to_owned()).await?;
- 
+
         Ok(serde_json::from_str(&res).unwrap())
     }
 
     ///Deletes the anime with `id` from the user's anime list
     ///
-    ///Returns 404 if the id isn't in the list. 
+    ///Returns 404 if the id isn't in the list.
     pub async fn delete_anime_list_item(&self, id: u32) -> Result<(), String> {
         let url = format!("https://api.myanimelist.net/v2/anime/{}/my_list_status", id);
-        let res = self.client.delete(url).bearer_auth(&self.access_token).send().await;
+        let res = self
+            .client
+            .delete(url)
+            .bearer_auth(&self.access_token)
+            .send()
+            .await;
         match res {
             Ok(r) => {
                 if r.status() == StatusCode::NOT_FOUND {
                     Err(format!("Anime {} not found", id))
-                }else {
+                } else {
                     Ok(())
                 }
             }
-            Err(e) => Err(format!("{}", e))
+            Err(e) => Err(format!("{}", e)),
         }
-       
     }
 
     //--Forum functions--//
-    
+
     ///Returns a vector of `HashMap`s that represent all the forum boards on MAL
     pub async fn get_forum_boards(&self) -> Result<ForumBoards, String> {
-        let res = self.do_request("https://api.myanimelist.net/v2/forum/boards".to_owned()).await?;
+        let res = self
+            .do_request("https://api.myanimelist.net/v2/forum/boards".to_owned())
+            .await?;
         self.parse_response(&res)
     }
 
-    
+    pub async fn get_forum_topic_detail(
+        &self,
+        topic_id: u32,
+        limit: Option<u8>,
+    ) -> Result<TopicDetails, String> {
+        let url = format!(
+            "https://api.myanimelist.net/v2/forum/topic/{}?limit={}",
+            topic_id,
+            limit.unwrap_or(100)
+        );
+        let res = self.do_request(url).await?;
+        self.parse_response(&res)
+    }
+
+    pub async fn get_forum_topics(
+        &self,
+        board_id: Option<u32>,
+        subboard_id: Option<u32>,
+        query: Option<String>,
+        topic_user_name: Option<String>,
+        user_name: Option<String>,
+        limit: Option<u32>,
+    ) -> Result<ForumTopics, String> {
+        let params = {
+            let mut tmp = vec![];
+            if let Some(bid) = board_id {
+                tmp.push(format!("board_id={}", bid.to_string()));
+            }
+            if let Some(bid) = subboard_id {
+                tmp.push(format!("subboard_id={}", bid.to_string()));
+            }
+            if let Some(bid) = query {
+                tmp.push(format!("q={}", bid.to_string()));
+            }
+            if let Some(bid) = topic_user_name {
+                tmp.push(format!("topic_user_name={}", bid.to_string()));
+            }
+            if let Some(bid) = user_name {
+                tmp.push(format!("user_name={}", bid));
+            }
+            tmp.push(format!("limit={}", limit.unwrap_or(100)));
+            tmp.join(",")
+        };
+        let url = format!("https://api.myanimelist.net/v2/forum/topics?{}", params);
+        let res = self.do_request(url).await?;
+        self.parse_response(&res)
+    }
 }
 
 #[derive(Deserialize, Debug)]